refactor(auth): 移除冗余的 config.ts，直接使用统一环境配置
- 删除 src/core/auth/config.ts
- oauth.ts 直接使用 getEnvironmentConfig 和 EnvConfig
- 简化代码结构，减少不必要的封装层
